Fix multi-user session handling, large index support, and back-navigation state

+++ CombinationGeneratorAPI/CombinationGeneratorAPI.Application/Interfaces/ICombinationService.cs
@@ -3,7 +3,7 @@ namespace CombinationGeneratorAPI.Application.Interfaces;
 public interface ICombinationService
 {
     long GetTotalCount(int n);
-    (int[] permutation, bool hasMore) GetNext(int n, int currentIndex);
-    IEnumerable<(int[] permutation, int index)> GetPage(int n, int fromIndex, int pageSize);
+    (int[] permutation, bool hasMore) GetNext(int n, long currentIndex);
+    IEnumerable<(int[] permutation, long index)> GetPage(int n, long fromIndex, int pageSize);
     int[] ComputePermutationAtIndex(int n, long index);
 }
